🔧fix: member_id 자료형 변경 (Integer -> Long)

+++ src/main/java/efub/SweetMeback/domain/member/oauth/service/OAuthService.java
@@ -17,6 +17,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
 
 import java.io.*;
+import java.math.BigInteger;
 import java.net.HttpURLConnection;
 import java.net.URL;
 
@@ -187,7 +188,7 @@ public class OAuthService{
     public Member getCurrentMember() {
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
         String principalName = authentication.getName();
-        Integer memberId = Integer.parseInt(principalName);
+        Long memberId = Long.parseLong(principalName);
         return memberRepository.findById(memberId).orElseThrow(() ->
                 new ResponseStatusException(HttpStatus.FORBIDDEN, "인증된 사용자 정보가 없습니다."));
     }
